jpegxl/interop-2026/features/bitdepth-10bpc.html

<!DOCTYPE html>
<title>JPEG XL interop: 10-bit decode</title>
<link rel="help" href="https://github.com/web-platform-tests/interop-jpegxl">
<meta name="assert" content="Decodes known 10-bit JPEG XL fixture (sunset_logo).">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/interop-helpers.js"></script>
<script>

promise_test(async () => {
  const img = await loadImage('../resources/conformance_sunset_logo.jxl');
  assert_equals(img.naturalWidth, 924);
  assert_equals(img.naturalHeight, 1386);
}, '10-bit JPEG XL fixture decodes and codestream orientation is applied.');

</script>
